add file 0

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -5,16 +5,17 @@
 /**
  * print_listint - prints all the elements of a listint_t
  * @h: The node to be printed
+ *
  * Return: The number of nodes
  */
 
 size_t print_listint(const listint_t *h)
 {
-	int n = 0;
+	size_t n = 0;
 
 	for (; h != NULL; n++)
 	{
-		printf("%u\n", h->n);
+		printf("%d\n", h->n);
 		h = h->next;
 	}
